create hook for game state
